feat:estructura ordenada y uniformizada

+++ src/application/use-cases/categoria-mueble/delete-categoria-mueble.use-case.ts
@@ -1,15 +1,20 @@
 import { inject, injectable } from "tsyringe";
+import { CategoriaMuebleException } from "../../../domain/exceptions/categoria-mueble.exception";
 import type { ICategoriaMuebleRepository } from "../../../domain/interfaces/categoria-mueble.repository.interface";
 import { DI_TOKENS } from "../../../common/IoC/tokens";
 
 @injectable()
 export class DeleteCategoriaMuebleUseCase {
-    constructor(
-        @inject(DI_TOKENS.ICategoriaMuebleRepository)
-        private readonly categoriaMuebleRepository: ICategoriaMuebleRepository,
-    ) {}
+  constructor(
+    @inject(DI_TOKENS.ICategoriaMuebleRepository)
+    private readonly categoriaMuebleRepository: ICategoriaMuebleRepository,
+  ) {}
 
-    async execute(id: string): Promise<void> {
-        await this.categoriaMuebleRepository.delete(id);
+  async execute(id: string): Promise<void> {
+    const existing = await this.categoriaMuebleRepository.findById(id);
+    if (!existing) {
+      throw CategoriaMuebleException.notFoundById(id);
     }
+    await this.categoriaMuebleRepository.delete(id);
+  }
 }
